Cellular

Cellular.jl

Cellular — Module.

Cellular provides a framework for building grid based simulations. Everything can be customised and added to, but there are some central idea that define how a Cellular simulation works: models, rules and neighborhoods. For input and output of data their are init arrays and outputs.

Models hold the configuration for a simulation, and trigger a specific rule method that operates on each of the cells in the grid. See AbstractModel and rule. Rules often trigger neighbors methods that sum surrounding cell neighborhoods (AbstractNeighborhood), such as Moore and Von Neumann neighborhoods.

Outputs are ways of storing of viewing the simulation, and can be used interchangeably depending on your needs. See AbstractOutput.

The inititialisation array may be any AbstractArray, containing whatever initialisation data is required to start the simulation. Most rules work on two-dimensional arrays, but one-dimensional arrays are also use for some cellular automata.

A typical simulation is run with a script like:

init = my_array
model = Life()
output = REPLOutput(init)

sim!(output, model, init)

Multiple models can be passed to sim!() in a tuple, and each of their rules will be run for the whole grid in sequence.

sim!(output, (model1, model2), init)

source

Cellular.jl

Example

Models and rules

Neighborhoods

Simulations

Output

Overflow

Customisation

Framework

Example

This example runs a game of life simulation, and uses the REPLOutput to print the frame directly in the REPL. You could instead use output = GtkOutput(init) for animation.

using Cellular

# Build a random starting grid
init = round.(Int8, max.(0.0, rand(-2.0:0.1:1.0, 70,70)))

# Use the default game of life model
model = Life()

# Use an output that shows the cellular automata in the REPL
output = REPLOutput(init)

sim!(output, model, init; time=1:5)

[0;0H[1;0H[2K[2;0H[2K[3;0H[2K[4;0H[2K[5;0H[2K[6;0H[2K[7;0H[2K[8;0H[2K[9;0H[2K[10;0H[2K[11;0H[2K[12;0H[2K[13;0H[2K[14;0H[2K[15;0H[2K[16;0H[2K[17;0H[2K[18;0H[2K[19;0H[2K[20;0H[2K[21;0H[2K[22;0H[2K[23;0H[2K[24;0H[2K[25;0H[2K[26;0H[2K[27;0H[2K[28;0H[2K[29;0H[2K[30;0H[2K[31;0H[2K[32;0H[2K[33;0H[2K[34;0H[2K[35;0H[2K[36;0H[2K[37;0H[2K[38;0H[2K[39;0H[2K[40;0H[2K[41;0H[2K[42;0H[2K[43;0H[2K[44;0H[2K[45;0H[2K[46;0H[2K[47;0H[2K[48;0H[2K[49;0H[2K[50;0H[2K[0;0H[1;0H[2K[2;0H[2K[3;0H[2K[4;0H[2K[5;0H[2K[6;0H[2K[7;0H[2K[8;0H[2K[9;0H[2K[10;0H[2K[11;0H[2K[12;0H[2K[13;0H[2K[14;0H[2K[15;0H[2K[16;0H[2K[17;0H[2K[18;0H[2K[19;0H[2K[20;0H[2K[21;0H[2K[22;0H[2K[23;0H[2K[24;0H[2K[25;0H[2K[26;0H[2K[27;0H[2K[28;0H[2K[29;0H[2K[30;0H[2K[31;0H[2K[32;0H[2K[33;0H[2K[34;0H[2K[35;0H[2K[36;0H[2K[37;0H[2K[38;0H[2K[39;0H[2K[40;0H[2K[41;0H[2K[42;0H[2K[43;0H[2K[44;0H[2K[45;0H[2K[46;0H[2K[47;0H[2K[48;0H[2K[49;0H[2K[50;0H[2K[0;0H[1;0H[2K[2;0H[2K[3;0H[2K[4;0H[2K[5;0H[2K[6;0H[2K[7;0H[2K[8;0H[2K[9;0H[2K[10;0H[2K[11;0H[2K[12;0H[2K[13;0H[2K[14;0H[2K[15;0H[2K[16;0H[2K[17;0H[2K[18;0H[2K[19;0H[2K[20;0H[2K[21;0H[2K[22;0H[2K[23;0H[2K[24;0H[2K[25;0H[2K[26;0H[2K[27;0H[2K[28;0H[2K[29;0H[2K[30;0H[2K[31;0H[2K[32;0H[2K[33;0H[2K[34;0H[2K[35;0H[2K[36;0H[2K[37;0H[2K[38;0H[2K[39;0H[2K[40;0H[2K[41;0H[2K[42;0H[2K[43;0H[2K[44;0H[2K[45;0H[2K[46;0H[2K[47;0H[2K[48;0H[2K[49;0H[2K[50;0H[2K[0;0H[1;0H[2K[2;0H[2K[3;0H[2K[4;0H[2K[5;0H[2K[6;0H[2K[7;0H[2K[8;0H[2K[9;0H[2K[10;0H[2K[11;0H[2K[12;0H[2K[13;0H[2K[14;0H[2K[15;0H[2K[16;0H[2K[17;0H[2K[18;0H[2K[19;0H[2K[20;0H[2K[21;0H[2K[22;0H[2K[23;0H[2K[24;0H[2K[25;0H[2K[26;0H[2K[27;0H[2K[28;0H[2K[29;0H[2K[30;0H[2K[31;0H[2K[32;0H[2K[33;0H[2K[34;0H[2K[35;0H[2K[36;0H[2K[37;0H[2K[38;0H[2K[39;0H[2K[40;0H[2K[41;0H[2K[42;0H[2K[43;0H[2K[44;0H[2K[45;0H[2K[46;0H[2K[47;0H[2K[48;0H[2K[49;0H[2K[50;0H[2K[0;0H[1;0H[2K[2;0H[2K[3;0H[2K[4;0H[2K[5;0H[2K[6;0H[2K[7;0H[2K[8;0H[2K[9;0H[2K[10;0H[2K[11;0H[2K[12;0H[2K[13;0H[2K[14;0H[2K[15;0H[2K[16;0H[2K[17;0H[2K[18;0H[2K[19;0H[2K[20;0H[2K[21;0H[2K[22;0H[2K[23;0H[2K[24;0H[2K[25;0H[2K[26;0H[2K[27;0H[2K[28;0H[2K[29;0H[2K[30;0H[2K[31;0H[2K[32;0H[2K[33;0H[2K[34;0H[2K[35;0H[2K[36;0H[2K[37;0H[2K[38;0H[2K[39;0H[2K[40;0H[2K[41;0H[2K[42;0H[2K[43;0H[2K[44;0H[2K[45;0H[2K[46;0H[2K[47;0H[2K[48;0H[2K[49;0H[2K[50;0H[2KCellular.REPLOutput{Cellular.ArrayOutput{Array{Int8,2}}}

More life-like examples (gleaned from CellularAutomata.jl):

# Morley
sim!(output, Life(b=[3,6,8], s=[2,4,5]), init; pause=0.1)

# 2x2
sim!(output, Life(b=[3,6], s=[1,2,5]), init; pause=0.05)

# Dimoeba
init1 = round.(Int8, max.(0.0, rand(70,70)))
sim!(output, Life(b=[3,5,6,7,8], s=[5,6,7,8]), init1; pause=0.1)

## No death
sim!(output, Life(b=[3], s=[0,1,2,3,4,5,6,7,8]), init; pause=0.1)

## 34 life
sim!(output, Life(b=[3,4], s=[3,4]), init; pause=0.1)

# Replicator
init2 = round.(Int8, max.(0.0, rand(70,70)))
init2[:, 1:30] .= 0
init2[21:50, :] .= 0
sim!(output, Life(b=[1,3,5,7], s=[1,3,5,7]), init2; pause=0.1)

Models and rules

Models define simulation behaviour. They hold data relevant to the simulation, and trigger dispatch of particular rule methods. Models can be chained together arbitrarily to make composite simulations.

Types and Constructors

Cellular.AbstractModel — Type.

abstract AbstractModel

A model contains all the information required to run a rule in a cellular simulation, given an initialised array. Models can be chained together in any order.

The output of the rule for an AbstractModel is written to the current cell in the grid.

source

Cellular.AbstractPartialModel — Type.

abstract AbstractPartialModel

An abstract type for models that do not write to every cell of the grid, for efficiency.

There are two main differences with AbstractModel. AbstractPartialModel requires initialisation of the destination array before each timestep, and the output of the rule is not written to the grid but done manually.

source

Cellular.AbstractLife — Type.

abstract AbstractLife <: Cellular.AbstractModel

Triggers dispatch on rule for game of life simulations. Models that extend this should replicate the fields for Life.

source

Cellular.Life — Type.

Game-of-life style cellular automata.

source

Methods

Cellular.rule — Function.

function rule(model, state, index, t, source, dest, args...)

Rules alter cell values based on their current state and other cells, often neighbors. Most rules return a value to be written to the current cell, except rules for models inheriting from AbstractPartialModel. These must write to the dest array directly.

Arguments:

model : AbstractModel

state: the value of the current cell

index: a (row, column) tuple of Int for the current cell coordinates

t: the current time step

source: the whole source array. Not to be written to

dest: the whole destination array. To be written to for AbstractPartialModel.

args: additional arguments passed through from user input to sim!

source

rule(model::AbstractLife, state, args...)

Rule for game-of-life style cellular automata.

The cell becomes active if it is empty and the number of neightbors is a number in the b array, and remains active the cell is active and the number of neightbors is in the s array.

Returns: boolean

source

Neighborhoods

Neighborhoods define a pattern of cells surrounding the current cell, and how they are combined to update the value of the current cell.

Types and Constructors

Cellular.AbstractNeighborhood — Type.

abstract AbstractNeighborhood

Abstract type to extend to a neighborhood

source

Cellular.AbstractRadialNeighborhood — Type.

abstract AbstractRadialNeighborhood{T} <: Cellular.AbstractNeighborhood

Abstract type to extend RadialNeighborhoods

source

Cellular.RadialNeighborhood — Type.

struct RadialNeighborhood{T, O} <: Cellular.AbstractRadialNeighborhood{T}

Radial neighborhoods calculate the surrounding neighborood from the radius around the central cell, with a number of variants.

They can be constructed with: RadialNeighborhood{:moore,Skip}(1,Skip()) but the keyword constructor should be preferable.

radius

The 'radius' of the neighborhood is the distance to the edge from the center cell. A neighborhood with radius 1 is 3 cells wide.

overflow

AbstractOverflow. Determines how co-ordinates outside of the grid are handled

source

Cellular.RadialNeighborhood — Method.

RadialNeighborhood(;typ = :moore, radius = 1, overflow = Skip)

The radial neighborhood constructor with defaults.

This neighborhood can be used for one-dimensional, Moore, von Neumann or Rotated von Neumann neigborhoods, and may have a radius of any integer size.

Keyword Arguments

typ : A Symbol from :onedim, :moore, :vonneumann or :rotvonneumann. Default: :moore

radius: Int. Default: 1

overflow: AbstractOverflow. Default: Skip()

source

Cellular.AbstractCustomNeighborhood — Type.

abstract AbstractCustomNeighborhood <: Cellular.AbstractNeighborhood

Custom neighborhoods are tuples of custom coordinates in relation to the central point of the current cell. They can be any arbitrary shape or size.

source

Cellular.CustomNeighborhood — Type.

struct CustomNeighborhood{H, O} <: Cellular.AbstractCustomNeighborhood

Allows completely arbitrary neighborhood shapes by specifying each coordinate specifically.

neighbors

A tuple of tuples of Int (or an array of arrays of Int, etc), contains 2-D coordinates relative to the central point

overflow

AbstractOverflow. Determines how co-ordinates outside of the grid are handled

source

Cellular.MultiCustomNeighborhood — Type.

struct MultiCustomNeighborhood{H, O} <: Cellular.AbstractCustomNeighborhood

Sets of custom neighborhoods that can have separate rules for each set.

multineighbors

A tuple of tuple of tuples of Int (or an array of arrays of arrays of Int, etc), contains 2-D coordinates relative to the central point.

cc

A vector the length of the base multineighbors tuple, for intermediate storage

overflow

AbstractOverflow. Determines how co-ordinates outside of the grid are handled

source

Methods

Cellular.neighbors — Function.

neighbors(hood::AbstractNeighborhood, state, index, t, source, args...)

Checks all cells in neighborhood and combines them according to the particular neighborhood type.

source

neighbors(hood::AbstractRadialNeighborhood{:onedim}, state, index, t, source, args...)

Sums single dimension radial neighborhoods. Commonly used by Wolfram.

source

neighbors(hood::AbstractRadialNeighborhood, state, index, t, source, args...)

Sums 2-dimensional radial Nieghborhoods. Specific shapes like Moore and Von Neumann are determined by inhood, as this method is general.

source

neighbors(hood::AbstractCustomNeighborhood, state, index, t, source, args...)

Sum a single custom neighborhood.

source

neighbors(hood::MultiCustomNeighborhood, state, index, t, source, args...)

Sum multiple custom neighborhoods separately.

source

Cellular.inhood — Function.

inhood(n::AbstractRadialNeighborhood{:moore}, p, q, row, col)

Check cell is inside a Moore neighborhood. Always returns true.

source

inhood(n::AbstractRadialNeighborhood{:vonneumann}, p, q, row, col)

Check cell is inside a Vonn-Neumann neighborhood, returning a boolean.

source

inhood(n::AbstractRadialNeighborhood{:rotvonneumann}, p, q, row, col)

Check cell is inside a Rotated Von-Neumann neighborhood, returning a boolean.

source

Simulations

Cellular.sim! — Function.

sim!(output, model, init, args...; time=1:1000, pause=0.0)

Runs the whole simulation, passing the destination aray to the passed in output for each time-step.

Arguments

output: An AbstractOutput to store frames or display them on the screen.

model: A single AbstractModel or a tuple of models that will each be run in sequence.

init: The initialisation array.

args: Any additional user defined args are passed through to rule and neighbors methods.

Keyword Arguments

time: Any Iterable of Number. Default: 1:1000

pause: A Number that specifies the pause beteen frames in seconds. Default: 0.0

source

Output

Types and Constructors

Cellular.AbstractOutput — Type.

abstract AbstractOutput

Simulation outputs are decoupled from simulation behaviour and can be used interchangeably. These outputs inherit from AbstractOutput.

Types that extend AbstractOutput define their own method for update_output.

source

Cellular.AbstractArrayOutput — Type.

abstract AbstractArrayOutput <: Cellular.AbstractOutput

Abstract type parent for array outputs.

source

Cellular.ArrayOutput — Type.

struct ArrayOutput{A} <: Cellular.AbstractArrayOutput

A simple array output that stores each step of the simulation in an array of arrays.

frames

An array that holds each frame of the simulation

source

Cellular.ArrayOutput — Method.

ArrayOutput(init)

Constructor for ArrayOutput

Arguments

init : the initialisation array

source

Cellular.REPLOutput — Type.

struct REPLOutput{A} <: Cellular.AbstractArrayOutput

A wrapper for ArrayOutput that is displayed as asccii blocks in the REPL.

array_output

source

Cellular.REPLOutput — Method.

REPLOutput(init)

Constructor for REPLOutput

Arguments

init: The initialisation array

source

Cellular.GtkOutput — Type.

struct GtkOutput{W, C, D} <: Cellular.AbstractOutput

Plot output live to a Gtk window.

window

canvas

scaling

ok

source

Cellular.GtkOutput — Method.

GtkOutput(init; scaling = 2)

Constructor for GtkOutput.

init::AbstractArray: the same init array that will also be passed to sim!()

source

Cellular.PlotsOutput — Type.

struct PlotsOutput{P} <: Cellular.AbstractOutput

A Plots.jl Output to plot cells as a heatmap in any Plots backend. Some backends (such as plotly) may be very slow to refresh. Others like gr() should be fine. using Plots must be called for this to be available.

plot

source

Cellular.PlotsOutput — Method.

PlotsOutput(init)

Constructor for GtkOutput.

init::AbstractArray: the init array that will also be passed to sim!()

source

Methods

Cellular.update_output — Function.

update_output(output, frame, t, pause)

Methods that update the output with the current frame, for timestep t.

source

update_output(output::AbstractArrayOutput, frame, t, pause)

Copies the current frame unchanged to the storage array

source

update_output(output::REPLOutput, frame, t, pause)

Extends update_output from ArrayOuput by also printing to the REPL.

source

update_output(output::GtkOutput, frame, t, pause)

Send current frame to the canvas in a Gtk window.

source

update_output(output::PlotsOutput, frame, t, pause)

source

Overflow

Your grids have edges. When neighborhood or spotting activities overflow past edge, you need a rule for deciding what to do.

Cellular.AbstractOverflow — Type.

abstract AbstractOverflow

Singleton types for choosing the grid overflow rule used in inbounds. These determine what is done when a neighborhood or jump extends outside of the grid.

source

Cellular.Wrap — Type.

struct Wrap <: Cellular.AbstractOverflow

Wrap cords that overflow to the opposite side

source

Cellular.Skip — Type.

struct Skip <: Cellular.AbstractOverflow

Skip coords that overflow boundaries

source

Customisation

Functions listed can be overridden or have methods added to them to modify simulation behaviour. Preferably create your own types of Model, Neighborhood or Output, and add methods for functions that dispatch on those types. If your new times and methods add useful functionality, consider making it publicly available by making pull request, or creating your own package that depends on Cellular.jl.

These are some more low-level functions you may also want to understand, if not actually modify extend.

Framework

Cellular.broadcast_rules! — Function.

broadcast_rules!(models, source, dest, index, t, args...)

Runs the rule(s) for each cell in the grid, dependin on the model(s) passed in. For [AbstractModel] the returned values are written to the dest grid, while for AbstractPartialModel the grid is pre-initialised to zero and rules manually populate the dest grid.

Returns a tuple containing the source and dest arrays for the next iteration.

source

Cellular.inbounds — Function.

inbounds(x, max, overflow)

Check grid boundaries for a single coordinate and max value or a tuple of coorinates and max values.

Returns a tuple containing the coordinate(s) followed by a boolean true if the cell is in bounds, false if not.

Overflow of type Skip returns the coordinate and false to skip coordinates that overflow outside of the grid. Wrap returns a tuple with the current position or it's wrapped equivalent, and true as it is allways in-bounds.

source

Cellular.process_image — Function.

process_image(frame, output)

Converts an array to an image format.

source