Introduction

Cellular.jl

Cellular — Module.

Cellular provides a framework for building grid based simulations. Everything can be customised, but there are a few central idea that define how a Cellular simulation works: models, output, and init arrays.

The typical simulation is run with the command:

model = Life()

sim!(output, model, init)

Multiple models can be passed to sim!() in a tuple.

sim!(output, (model1, model2), init)

The init array may be any AbstractArray, containing some whatever initialisation data is required. Most rules two-dimensional arrays, but one dimensional arrays are also use for some Cellular automata. model and outputs can be types defined in Cellular, in packages that extend Cellular, or custom types.

Exported types and methods

AbstractArrayOutput

AbstractCustomNeighborhood

AbstractInPlaceModel

AbstractLife

AbstractModel

AbstractNeighborhood

AbstractOutput

AbstractOverflow

AbstractRadialNeighborhood

ArrayOutput

CustomNeighborhood

Life

MultiCustomNeighborhood

REPLOutput

RadialNeighborhood

Skip

Wrap

automate!

inbounds

show

sim!

source

Example

This example runs a game of life simulation, and uses the REPLOutput to print the final frame directly in the REPL. You could intsead use output = TkOutput(init) for live animation.

using Cellular

# Build a random starting grid
init = round.(Int8, max.(0.0, rand(-2.0:0.1:1.0, 30,70)))

# Use the default game of life model
model = Life()

# Use at output that shows life in the REPL
output = REPLOutput(init)

sim!(output, model, init; time=1:5)

Cellular.REPLOutput{Array{Int8,2}}

These are some more life-like examples (gleaned from CellularAutomata.jl)

# Morley
sim!(output, Life(B=[3,6,8],S=[2,4,5]), init)

# 2x2
sim!(output, Life(B=[3,6],S=[1,2,5]), init)

# Replicator
sim!(output, Life(B=[1,3,5,7], S=[1,3,5,7]), init)

# Dimoeba
sim!(output, Life(B=[3,5,6,7,8], S=[5,6,7,8]), init)

## No death
sim!(output, Life(B=[3], S=[0,1,2,3,4,5,6,7,8]), init)

## 34 life
sim!(output, Life(B=[3,4], S=[3,4]), init)

Models and rules

Models define modelling behaviour. They hold data relevant to the simulation, and trigger dispatch of particular rule methods. Models can be chained together arbitrarily to make composite simulations.

Cellular.AbstractModel — Type.

A module contains all the data required to run a rule in a cellular simulation. Models can be chained together in any order or length.

The output of the rule for a default modules is written to the current cell in the grid.

source

Cellular.AbstractInPlaceModel — Type.

The output of In-place modules is ignored, instead they manually update cells. This is the best options for modules that only update a subset of cells.

source

Cellular.AbstractLife — Type.

abstract AbstractLife <: Cellular.AbstractModel

Triggers dispatch on rule for game of life simulations. Models that extend this should replicate the fields for Life.

source

Cellular.Life — Type.

struct Life{N, B, S} <: Cellular.AbstractLife

Game of life style cellular automata.

Fields:

neighborhood::AbstractNeighborhood: The default is a :moore RadialNeighborhood with Wrap overflow

b: Array or Tuple of integers to match neighbors when cell is empty Default = [3]

s: Array, Tuple or Iterable of integers to match neighbors when cell is full. The default is [2,3]

source

Neighborhoods

Some rules require neighborhoods as a field of the model. Neighborhoods define the pattern of cells that surrounds the current cell, and how they are summed to set the value of the current cell in a model. Neighborhoods are generally not used in AbstractInPlaceModel.

Cellular.AbstractNeighborhood — Type.

abstract AbstractNeighborhood

Neighborhoods define the behaviour towards the cells surrounding the current cell.

source

Cellular.AbstractRadialNeighborhood — Type.

abstract AbstractRadialNeighborhood{T} <: Cellular.AbstractNeighborhood

Radial neighborhoods calculate the neighborood in a loop from simple rules base of the radius of cells around the central cell.

source

Cellular.RadialNeighborhood — Type.

struct RadialNeighborhood{T, O} <: Cellular.AbstractRadialNeighborhood{T}

radius

overflow

source

Cellular.AbstractCustomNeighborhood — Type.

Custom neighborhoods are tuples of custom coordinates in relation to the central point of the current cell. They can be any arbitrary shape or size.

abstract AbstractCustomNeighborhood <: Cellular.AbstractNeighborhood

source

Cellular.CustomNeighborhood — Type.

struct CustomNeighborhood{H, O} <: Cellular.AbstractCustomNeighborhood

neighbors

overflow

source

Cellular.MultiCustomNeighborhood — Type.

Multi custom neighborhoods are sets of custom neighborhoods that can have separate rules for each set. cc is a vector used to store the output of these rules.

struct MultiCustomNeighborhood{H, O} <: Cellular.AbstractCustomNeighborhood

multineighbors

cc

overflow

source

Overflow

Your grids have edges. When neighborhood or spotting activities overflow past edge, you need a rule for deciding what to do.

Cellular.AbstractOverflow — Type.

Singleton for selection overflow rules. These determine what is done when a neighborhood or jump extends outside the grid.

source

Cellular.Wrap — Type.

Wrap()

Wrap cords that overflow to the opposite side

source

Cellular.Skip — Type.

Skip()

Skip coords that overflow boundaries

source

Simulations

Cellular.sim! — Function.

sim!(output, model, init, args; time, pause)

Runs the whole simulation, passing the destination aray to the passed in output for each time-step.

Arguments

output: Any AbstractOutput to save frames to or display on the screen.

model: A single mode (AbstractModel) or a tuple of models.

init: Initialisation array.

Keyword Arguments

time: Any Iterable of Number. Default is 1:1000

pause: A Number, pauses beteen frames. Default is 0.0.

source

Output

Cellular.AbstractOutput — Type.

abstract AbstractOutput

Simulation outputs can be used interchangeably as they are decoupled from the simulation behaviour. Outputs should inherit from AbstractOutput.

All types extending AbstractOutput should have their own method of update_output.

source

Cellular.AbstractArrayOutput — Type.

abstract AbstractArrayOutput <: Cellular.AbstractOutput

Output subtype for arrays

source

Cellular.ArrayOutput — Type.

struct ArrayOutput{A} <: Cellular.AbstractArrayOutput

Simple array output: creates an array of frames.

frames

source

Cellular.REPLOutput — Type.

struct REPLOutput{A} <: Cellular.AbstractArrayOutput

An array output that is printed as asccii blocks in the REPL.

frames

source

Custom show() methods are available for some outputs.

Base.show — Function.

show(io, output)

Print the last frame of a simulation in the REPL.

source

Customisation

These functions (and those already listed) can all be overridden to change simulation behaviour. Preferably create your own type of Model, Neighborhood or Output, and add methods of these functions that dispatch on those types. If your new times and methods add broadly useful functionality, consider making it publicly available by making pull request, or creating your own package that depends on Cellular.jl.

Framework

Cellular.rule — Function.

rule(model, state, index, t, source, args)

Rules for altering cell values

Arguments:

rule::AbstractModel:

state: value of the current cell

index: row, column coordinate tuple for the current cell

t: current time step

source: the whole source array

args: additional arguments passed through from user input to sim!

source

rule(model::AbstractLife, state, args...) = begin

Rule for game-of-life style cellular automata.

Cell value is flipped if cell is empty and the bumber of neightbors is in the b array, or if the cell is full and the bumber of neightbors is in the s array.

Only the model and state arguments are used.

Returns: boolean.

source

Cellular.broadcastrules! — Function.

broadcastrules!(models::Tuple{T,Vararg}, source, dest, index, t, args...) where {T<:AbstractModel}

Broadcast rule over each cell in the grid, for each module. Returned values are written to the dest grid.

source

broadcastrules!(models::Tuple{T,Vararg}, source, dest, index, t, args...) where {T<:AbstractInPlaceModel}

AbstractInPlaceModel Broadcasts rules for each cell in the grid, for each module. Rules must manually write to the source array. return values are ignored.

source

Cellular.automate! — Function.

automate!(models::Tuple, source, dest, t, args...) = begin

Runs the rules over the whole grid, for each module in sequence.

source

Neighborhoods

Cellular.neighbors — Function.

neighbors(h::AbstractNeighborhood, state, index, t, source, args...) = begin

Checks all cells in neighborhood and combines them according to the particular neighborhood rule.

neighbors(h, state, index, t, source, args)

defined at Cellular/src/neighborhoods.jl:72.

neighbors(h, state, index, t, source, args)

defined at Cellular/src/neighborhoods.jl:74.

neighbors(h, state, index, t, source, args)

defined at Cellular/src/neighborhoods.jl:85.

neighbors(h, state, index, t, source, args)

defined at Cellular/src/neighborhoods.jl:101.

neighbors(h, state, index, t, source, args)

defined at Cellular/src/neighborhoods.jl:104.

source

Cellular.inbounds — Function.

inbounds(xs::Tuple, maxs::Tuple, overflow)

Check grid boundaries for two coordinates.

Returns a 3-tuple of coords and a boolean. True means the cell is in bounds, false it is not.

source

inbounds(x::Number, max::Number, overflow::Skip)

Skip coordinates that overflow outside of the grid.

source

inbounds(x::Number, max::Number, overflow::Skip)

Swap overflowing coordinates to the other side.

source

Cellular.inhood — Function.

inhood(n::AbstractRadialNeighborhood{T}, p, q, row, col)

Check cell is inside a radial neighborhood, returning a boolean.

source

Output

Cellular.update_output — Function.

update_output(output, frame, t, pause)

Methods that update the output with the current frame, for timestep t.

update_output(output, frame, t, pause)

defined at Cellular/src/output.jl:39.

update_output(output, frame, t, pause)

defined at Cellular/src/output.jl:55.

source

Cellular.process_image — Function.

process_image(frame, output)

Converts an array to an image format.

source