CellularAutomataBase.jl

CellularAutomataBase.jl

CellularAutomataBase — Module.

CellularAutomataBase provides a framework for building grid-based simulations.

The framework is highly customisable, but there are some central ideas that define how a simulation works: rules, init arrays and outputs.

Rules hold the configuration for a simulation, and trigger a specific applyrule method that operates on each of the cells in the grid. See AbstractRule and applyrule. Rules come in a number of flavours, which allows assumptions to be made about running them that can greatly improve performance. Rules are chained together in a Ruleset object.

The init array may be any AbstractArray, containing whatever initialisation data is required to start the simulation. The Array type and element type of the init array determine the types used in the simulation, as well as providing the initial conditions.

Outputs are ways of storing of viewing a simulation, and can be used interchangeably depending on your needs. See AbstractOutput.

A typical simulation is run with a script like:

init = my_array
rules = Ruleset(Life())
output = ArrayOutput(init)

sim!(output, rules; init=init)

Multiple models can be passed to sim! in a Ruleset. Each rule will be run for the whole grid, in sequence.

sim!(output, Ruleset(rule1, rule2); init=init)

For better performance, models included in a tuple will be combined into a single model (with only one array write). This is limited to AbstractCellRule, although AbstractNeighborhoodRule may be used as the first model in the tuple.

sim!(output, Rules(rule1, (rule2, rule3)); init=init)

source

Examples

While this package isn't designed or optimised specifically to run the game of life, it's a simple example of what this package can do. This example runs a game of life and displays it in a REPLOutput.

using CellularAutomataBase

# Build a random starting grid
init = round.(Int8, max.(0.0, rand(-2.0:0.1:1.0, 70,70)))

# Use the default game of life model
model = Ruleset(Life())

# Use an output that shows the cellular automata as blocks in the REPL
output = REPLOutput{:block}(init; fps=100)

sim!(output, model, init; tstop=5)

More life-like examples:

# Morley
sim!(output, Ruleset(Life(b=[3,6,8], s=[2,4,5]); init=init))

# 2x2
sim!(output, Ruleset(Life(b=[3,6], s=[1,2,5]); init=init))

# Dimoeba
init1 = round.(Int8, max.(0.0, rand(70,70)))
sim!(output, Ruleset(Life(b=[3,5,6,7,8], s=[5,6,7,8]); init=init1))

## No death
sim!(output, Ruleset(Life(b=[3], s=[0,1,2,3,4,5,6,7,8]); init))

## 34 life
sim!(output, Ruleset(Life(b=[3,4], s=[3,4])); init=init, fps=10)

# Replicator
init2 = round.(Int8, max.(0.0, rand(70,70)))
init2[:, 1:30] .= 0
init2[21:50, :] .= 0
sim!(output, Ruleset(Life(b=[1,3,5,7], s=[1,3,5,7])); init=init2)

Rules

Rules define simulation behaviour. They hold data relevant to the simulation, and trigger dispatch of particular applyrule or applyrule! methods. Rules can be chained together arbitrarily to make composite simulations.

Types and Constructors

CellularAutomataBase.AbstractRule — Type.

abstract type AbstractRule

A rule contains all the information required to run a rule in a cellular simulation, given an initial array. Rules can be chained together sequentially.

The output of the rule for an AbstractRule is allways written to the current cell in the grid.

source

CellularAutomataBase.AbstractCellRule — Type.

abstract type AbstractCellRule <: AbstractRule

A Rule that only writes and accesses a single cell: its return value is the new value of the cell. This limitation can be useful for performance optimisations.

Accessing the data.source and data.dest arrays directly is not guaranteed to have correct results, and should not be done.

source

CellularAutomataBase.AbstractNeighborhoodRule — Type.

abstract type AbstractNeighborhoodRule <: AbstractRule

A Rule That only accesses a neighborhood, defined by its radius distance from the current cell.

For each cell a buffer will be populated containing the neighborhood cells, accessible with buffer(data). This allows memory optimisations and the use of BLAS routines on the neighborhood. It also means that and no bounds checking is required.

AbstractNeighborhoodRule must read only from the state variable and the neighborhood_buffer array, and never manually write to the dest(data) array. Its return value is allways written to the central cell.

Custom Neighborhood rules must return their radius with a radius() method.

source

CellularAutomataBase.AbstractPartialRule — Type.

abstract type AbstractPartialRule <: AbstractRule

AbstractPartialRule is for rules that manually write to whichever cells of the grid that they choose, instead of updating every cell with their output.

Updates to the destination array (dest(data)) must be performed manually, while the source array can be accessed with source(data).

The dest array is copied from the source prior to running the applyrule! method.

source

CellularAutomataBase.AbstractPartialNeighborhoodRule — Type.

abstract type AbstractPartialNeighborhoodRule <: AbstractPartialRule

A Rule that only writes to its neighborhood, defined by its radius distance from the current point. TODO: should this exist?

Custom PartialNeighborhood rules must return their radius with a radius() method.

source

CellularAutomataBase.Life — Type.

Rule for game-of-life style cellular automata.

Field	Description	Default	Limits

neighborhood	Any AbstractNeighborhood	RadialNeighborhood{1}()	nothing

b	Array, Tuple or Iterable of integers to match neighbors when cell is empty	(3, 3)	(0, 8)

s	Array, Tuple or Iterable of integers to match neighbors cell is full	(2, 3)	(0, 8)

source

Neighborhoods

Neighborhoods define a pattern of cells surrounding the current cell, and how they are combined to update the value of the current cell.

Types and Constructors

CellularAutomataBase.AbstractNeighborhood — Type.

abstract type AbstractNeighborhood

Neighborhoods define how surrounding cells are related to the current cell. The neighbors function returns the sum of surrounding cells, as defined by the neighborhood.

source

CellularAutomataBase.RadialNeighborhood — Type.

struct RadialNeighborhood{R} <: AbstractNeighborhood{R}

Radial neighborhoods calculate the surrounding neighborood from the radius around the central cell. The central cell is ommitted.

source

CellularAutomataBase.AbstractCustomNeighborhood — Type.

abstract type AbstractCustomNeighborhood <: AbstractNeighborhood{R}

Custom neighborhoods are tuples of custom coordinates in relation to the central point of the current cell. They can be any arbitrary shape or size.

source

CellularAutomataBase.CustomNeighborhood — Type.

Allows completely arbitrary neighborhood shapes by specifying each coordinate specifically.

source

CellularAutomataBase.LayeredCustomNeighborhood — Type.

Sets of custom neighborhoods that can have separate rules for each set.

source

Output

Output Types and Constructors

CellularAutomataBase.AbstractOutput — Type.

abstract type AbstractOutput <: AbstractArray{T,1}

All outputs must inherit from AbstractOutput.

Simulation outputs are decoupled from simulation behaviour and in many cases can be used interchangeably.

source

CellularAutomataBase.ArrayOutput — Type.

A simple output that stores each step of the simulation in a vector of arrays.

Arguments:

frames: Single init array or vector of arrays

tstop: The length of the output.

source

CellularAutomataBase.REPLOutput — Type.

An output that is displayed directly in the REPL. It can either store or discard simulation frames.

Arguments:

frames: Single init array or vector of arrays

Keyword Arguments:

fps::Real: frames per second to run at

showfps::Real: maximum displayed frames per second

store::Bool: store frames or not

color: a color from Crayons.jl

cutoff::Real: the cutoff point to display a full or empty cell. Default is 0.5

To choose the display type can pass :braile or :block to the constructor:

REPLOutput{:block}(init)

The default option is :block.

source

Frame processors

CellularAutomataBase.AbstractFrameProcessor — Type.

abstract type AbstractFrameProcessor

Frame processors convert frame data into RGB24 images They can be passed as procesor keyword argument to outputs that have an image display.

To add new processor, define a type that inherits from AbstractFrameProcessor and a frametoimage method:

frametoimage(p::YourProcessor, output, frame, t) = some_rbg_image

source

CellularAutomataBase.GreyscaleProcessor — Type.

struct GreyscaleProcessor <: AbstractFrameProcessor

Converts output frames to a greyscale images

source

CellularAutomataBase.GreyscaleZerosProcessor — Type.

struct GreyscaleZerosProcessor{C} <: AbstractFrameProcessor

" Converts output frames to a greyscale image with the chosen color for zeros. Usefull for separating low values from actual zeros

Arguments

zerocolor: RGB24 or a value that will be converted to RGB24 by the RGB24() constructor.

source

CellularAutomataBase.ColorSchemeProcessor — Type.

struct ColorSchemeProcessor{S} <: AbstractFrameProcessor

" Converts output frames to a colorsheme.

Arguments

scheme: a ColorSchemes.jl colorscheme.

source

CellularAutomataBase.ColorSchemeZerosProcessor — Type.

struct ColorSchemeZerosProcessor{S, C} <: AbstractFrameProcessor

" Converts frame to a colorshceme image with the chosen color for zeros. Usefull for separating low values from actual zeros

Arguments

scheme: a ColorSchemes.jl colorscheme. zerocolor: RGB24 or a value that will be converted to RGB24 by the RGB24() constructor.

source

Overflow

Missing docstring.

Missing docstring for AbstractOverflow. Check Documenter's build log for details.

CellularAutomataBase.WrapOverflow — Type.

struct WrapOverflow <: CellularAutomataBase.AbstractOverflow

Wrap cords that overflow boundaries back to the opposite side

source

CellularAutomataBase.RemoveOverflow — Type.

struct RemoveOverflow <: CellularAutomataBase.AbstractOverflow

Remove coords that overflow boundaries

source

Methods

CellularAutomataBase.distances — Method.

distances(a::AbstactMatrix)

Calculate the matrix of distances (in units of cells) between all cells in a matrix

source

CellularAutomataBase.replay — Method.

replay(output::AbstractOutput)

Show a stored simulation again. You can also use this to show a simulation in different output type.

If you ran a simulation with store=false there won't be much to replay.

Example

replay(REPLOutput(output))

source

CellularAutomataBase.resume! — Method.

resume!(output, ruleset; tstop=100)

Restart the simulation where you stopped last time.

Arguments

See sim!.

source

CellularAutomataBase.ruletypes — Method.

Return a tuple of the base types of the rules in the ruleset

source

CellularAutomataBase.savegif — Method.

savegif(filename::String, output::AbstractOutput)

Write the output array to a gif. Saving very large gifs may trigger a bug in imagemagick.

source

CellularAutomataBase.sim! — Method.

sim!(output, ruleset, init; tstop=1000)

Runs the whole simulation, passing the destination aray to the passed in output for each time-step.

Arguments

output: An AbstractOutput to store frames or display them on the screen.

ruleset: A Rule() containing one ore more AbstractRule. These will each be run in sequence.

init: The initialisation array.

args: additional args are passed through to rule and neighbors methods.

Keyword Arguments

tstop: Any Number. Default: 100

source

CellularAutomataBase.addpadding — Method.

Find the maximum radius required by all rules Add padding around the original init array, offset into the negative So that the first real cell is still 1, 1

source

CellularAutomataBase.applyrule — Function.

applyrule(rule::AbstractRule, data, state, index)

Updates cell values based on their current state and the state of other cells as defined in the Rule.

Arguments:

rule : AbstractRule

data : FrameData

state: the value of the current cell

index: a (row, column) tuple of Int for the current cell coordinates - t: the current time step

args: additional arguments passed through from user input to sim!

Returns a value to be written to the current cell.

source

CellularAutomataBase.applyrule! — Function.

applyrule!(rule::AbstractPartialRule, data, state, index)

A rule that manually writes to the dest array, used in rules inheriting from AbstractPartialRule.

Arguments:

see applyrule

source

CellularAutomataBase.applyrule — Method.

rule(model::AbstractLife, state)

Rule for game-of-life style cellular automata. This is a demonstration of Cellular Automata more than a seriously optimised game of life model.

Cells becomes active if it is empty and the number of neightbors is a number in the b array, and remains active the cell is active and the number of neightbors is in the s array.

Returns: boolean

Examples (gleaned from CellularAutomata.jl)

Use the arrow keys to scroll around, or zoom out if your terminal can do that!

# Life. 
init = round.(Int64, max.(0.0, rand(-3.0:0.1:1.0, 300,300)))
output = REPLOutput{:block}(init; fps=10, color=:red)
sim!(output, model, init; time=1000)

# Dimoeba
init = rand(0:1, 400, 300)
init[:, 100:200] .= 0
output = REPLOutput{:braile}(init; fps=25, color=:blue)
sim!(output, Ruleset(Life(b=(3,5,6,7,8), s=(5,6,7,8))), init; time=1000)

# Replicator
init = fill(1, 300,300)
init[:, 100:200] .= 0
init[10, :] .= 0
output = REPLOutput{:block}(init; fps=60, color=:yellow)
sim!(output, Ruleset(Life(b=(1,3,5,7), s=(1,3,5,7))), init; time=1000)

source

CellularAutomataBase.applyrule — Method.

applyrule(rules::Tuple, data, state, (i, j))

Subrules. If a tuple of rules is passed to applyrule, run them sequentially for each cell. This can have much beter performance as no writes occur between rules, and they are essentially compiled together into compound rules. This gives correct results only for AbstractCellRule, or for a single AbstractNeighborhoodRule followed by AbstractCellRule.

source

CellularAutomataBase.frametoimage — Function.

Convert frame matrix to RGB24, using any AbstractFrameProcessor

source

CellularAutomataBase.handleoverflow! — Method.

Wrap overflow where required. This optimisation allows us to ignore bounds checks on neighborhoods and still use a wraparound grid.

source

CellularAutomataBase.hoodsize — Method.

sizefromradius(radius)

Get the size of a neighborhood dimension from its radius, which is always 2r + 1.

source

CellularAutomataBase.inbounds — Method.

inbounds(x, max, overflow)

Check grid boundaries for a single coordinate and max value or a tuple of coorinates and max values.

Returns a tuple containing the coordinate(s) followed by a boolean true if the cell is in bounds, false if not.

Overflow of type RemoveOverflow returns the coordinate and false to skip coordinates that overflow outside of the grid. WrapOverflow returns a tuple with the current position or it's wrapped equivalent, and true as it is allways in-bounds.

source

CellularAutomataBase.initframes! — Method.

Frames are deleted and reallocated during the simulation, as performance is often display limited, and this allows runs of any length.

source

CellularAutomataBase.initframes! — Method.

Frames are preallocated and reused.

source

CellularAutomataBase.initstatus! — Method.

Initialise the block status array. This tracks whether anything has to be done in an area of the main array.

source

CellularAutomataBase.maprule! — Method.

Apply the rule for each cell in the grid, using optimisations allowed for the supertype of the rule.

source

CellularAutomataBase.maxradius — Method.

Find the largest radius present in the passed in rules.

source

CellularAutomataBase.neighbors — Function.

neighbors(hood::AbstractNeighborhood, state, indices, t, source, args...)

Checks all cells in neighborhood and combines them according to the particular neighborhood type.

source

CellularAutomataBase.neighbors — Method.

neighbors(hood::AbstractCustomNeighborhood, buf, state)

Sum a single custom neighborhood.

source

CellularAutomataBase.neighbors — Method.

neighbors(hood::LayeredCustomNeighborhood, buf, state) Sum multiple custom neighborhoods separately.

source

CellularAutomataBase.neighbors — Method.

neighbors(hood::RadialNeighborhood, buf, state)

Sums moore nieghborhoods of any dimension.

source

CellularAutomataBase.radius — Function.

Return the radius of a rule if it has one, otherwise zero.

source

CellularAutomataBase.runsim! — Method.

run the simulation either directly or asynchronously.

source

CellularAutomataBase.sequencerules! — Method.

Iterate over all rules recursively, swapping source and dest arrays. Returns a tuple containing the source and dest arrays for the next iteration.

source

CellularAutomataBase.showframe — Method.

showframe(output::AbstractOutput, [t])

Show the last frame of the output, or the frame at time t.

source

CellularAutomataBase.simdata — Method.

Generate simulation data to match a ruleset and init array.

source

CellularAutomataBase.simloop! — Method.

Loop over the selected timespan, running the ruleset and displaying the output

source

CellularAutomataBase.swapsource — Method.

Swap source and dest arrays. Allways returns regular SimData.

source

CellularAutomataBase.updatetime — Method.

Uptate timestamp

source